Doc: do not reference chapters from commands list
The commands list is shared between the HTML doc and the manpage, but
the manpage contains only a subset of HTML's chapters, so references
point nowhere. Removing the references is the easy fix; it shouldn't
affect usability of HTML all that much because the user has a menu of
topics, and will probably search for keywords to find where else the
topic is mentioned.

Fixes #1830.