CliArgsParser: change `argv` from `Vec<String>` to `Vec<OsString>` (#1839)
* Changed CliArgsParser::new() to take Vec<OsString> and added the BytesVec type
a CXX shared type, which is a wrapper around a Vec<u8> to hold the input
from the command line in an unopinionated way. Fixes issue https://github.com/newsboat/newsboat/issues/1357

Tests not changed in this commit.

* Changed CliArgsParser::new() to take Vec<OsString>
Also added the BytesVec type a CXX shared type,
which is a wrapper around a Vec<u8> to hold the input
from the command line in an unopinionated way.
Fixes issue https://github.com/newsboat/newsboat/issues/1357

Tests not changed in this commit.

* Changed test suits to account for issue https://github.com/newsboat/newsboat/issues/1357

Changed the tests, so that instead of to_string() calls
there's now into() calls, which is a bit more generic and also
allows further change (assuming &str impls Into<T> for that type).

All tests pass

* fixed formatting

* Fixed requested changes.
Conditional compilation removed.
Change c++ while loop to conform to idiom
Allow for lossy conversion of argv program name.