Replaced clap-rs with lexopt as requested by issue https://github.com/newsboat/newsboat/issues/1549
Added lexopt and required parsing loop.
Code is a bit more verbose, due to the nature of lexopt.
Added LexoptWrapper to be able to track last seen
option type, if it was in long form or short form.
This adds an additional check per option (not the values though).
The reason for this additional check was to reduce code bloat
and thus make it a lot easier to read, I figured this was a trade off
we're willing to do, since it's not in a time critical section of the
application.

Lexopt also does not come with built in functionality
for handling short options "-o=value", this
therefore also required some patching over,
by the local function strip_eq(). An issue
has been filed with the crate maintainer who has
agreed to implement the feature, so in the future
we can change this (and thus also remove the
"double checking" as mentioned above).