Replaced cargo lexopt dependency with custom version
Hacked on lexopt and added additional functions to the public
API of lexopt, called .key_value() which returns a new type
called ValueType which is an enum of:
Normal(OsString)
KVPValue(OsString)

.key_value() looks at current/next parsed value
and if it contains an '=' it returns a KVPValue(OsString)
where the equal sign is stripped, if it doesn't,
it returns a Normal value. Doing this, makes it so
we can actually check if it was a key value pair that we parsed
or not. Simple, easy solution, no involved extra parsing,
no running through all arguments.

Also; when calling .value() - the equal sign is _always_
stripped from the final value, just like it was before.