Added tests & Changed tests titled wrong
One test was titled "Supports `=` between combined short options and value"
when in fact it only tested for 1 short option and equal sign.

I've duplicated that test, changed the title of the old one
to "Supports `=` between short option and value"
and implemented the test that checks actual combined short options and value.