Make command line options bold/underline in man pages
In most other man pages, that I looked at, the program options are
formatted bold and their values are underlined. So let's do this, too,
to provide an overall consistent look.

Also for newsboat's `-x` option in the synopsis the angle brackets
around the `command` were removed to make it consistent to all the other
options. The space in front of the ellipsis was removed because in the
options section there was also no leading space. This also seems to be
the common style in most other man pages.