CI: limit greedy containers to 8 tasks (down from 10)
Greedy containers are a feature of Cirrus CI that lets us use more CPUs
than we requested. This is beneficial to us because our builds are
faster, and to Cirrus because we utilize the resources better. At the
moment, our tasks request 4 CPUs and run 10 Make tasks.

Recently, we started seeing CI failures with greedy tasks. The dashboard
indicates that we get 10 CPUs and run out of memory.

In an attempt to curb that, I lowered the number of Make tasks to eight.
Let's see how it goes.