Add possibility to jump to previous item when toggling an item as read
When an item is toggled as read, newsboat will jump to the next item in
the list. New config parameter is added which causes newsboat to jump to
the previous item in the list after an item is toggled as read.

This is useful when reading items from the oldest entry in the feed.