Do not convert feeds based on Content-Type
This reverts PRs #2214 and #2243, containing the following commits:

203bbf80206180740f24137ce0f35fcb64f2e56e
    Make utils::convert_text() available to C++
5dfdc5d64d79a2c080956acf599d202b4aa2988b
    Convert feed data to utf-8 if charset specifies non-"utf-8" encoding
c7fd68105d35b4c3c7e03c4013768272b6c3d015
    retrieve_url(): Convert data to utf-8 if non-utf8 charset is
    specified
7eb721fdfab39d64a455ec8c6d90681864e057c7
    Avoid ODR violation by defining HeaderValues in anonymous namespaces
7ba7470af1635e736f2e688fb46303e3569656f0
    Get rid of raw new/delete useage in parser.cpp's handle_headers
1c3e1618f780655375b465e6a1d7ef030cb4b460
    Reset HTTP headers when detecting a new response
64ad6de02d7f3fd8239772c70e964895360c7524
    Update rss/parser.cpp
4b0108a1a501ac69ca969c4790990c35148fd7b0
    Update rss/parser.cpp

These commits led to problems with feeds that specify the same encoding
both in the Content-Type and the XML tag; these were converted to UTF-8,
but then the XML parser tried to convert them to UTF-8 again. We don't
have a good fix for this yet. The problem is tracked in #1436.