Avoid calling virtual functions from destructor
This fixes a cppcheck warning about `~EnvVar` calling virtual `set()`
and `unset()`: in destructors, these calls would be resolved to
`EnvVar`'s implementations, not implementations from the derived class.
But we don't need to override those implementations anyway; the `EnvVar`
is structures in such a way that the only thing that changes in derived
classes is the `on_change_fn`. This also means that there is no need for
`on_change()` setter to be `virtual` either.