config.sh: fix misdetection of missing commands
The script used a bashism to redirect stdout and stderr to /dev/null:

    if cmd &> /dev/null ; then ...

With a POSIX shell such as dash, this line is interpreted as "run cmd in
background, redirecting its stdout to /dev/null". Since forking into
background always succeeds, this entire `if` always went into the main
branch and declared that the `cmd` exists.

This caused config.sh to think that Asciidoctor and Cargo are installed
even if they aren't.