config.sh: do not check for Asciidoctor
dcb6038ee0222274e4b9f02a10fbacb462d4765e introduced a check for commands
needed to build Newsboat: Asciidoctor and Cargo. However, Asciidoctor is
optional because it's only used for building docs (#1885). This commit
removes the check, so Newsboat can truly be built with just the
necessary dependencies installed.

This makes the build process slightly less nice, because the user won't
get an up-front warning about the missing Asciidoctor. To regain that
warning, I think we'd need to check for the program from within Makefile
iff `doc` target is being built (or any target that depends on `doc`).
This sounds too complicated to be worthwhile, though.

Fixes #2353.