Add --export-to-opml2 option to export including tags
OPML, while primarily using a hierarchy of outlines, also supports tags
in version 2.0 (http://opml.org/spec2.opml#1629042387000):

> category is a string of comma-separated slash-delimited category
> strings, in the format defined by the RSS 2.0 category element. To
> represent a "tag," the category string should contain no slashes.
> Examples: 1. category="/Boston/Weather".
> 2. category="/Harvard/Berkman,/Politics".

By exporting as OPML 2.0 we can include the tags in a way other programs
understand. Commas are replaced with underscores. Slashes are left
as-is, so when Newsboat imports a hierarchical OPML file and then
exports it again, that hierarchy isn't lost.

For backwards compatibility with software that may not understand OPML
2.0, introduce this functionality as a new option, --export-to-opml2.

Closes: #871