CI: add GCC 13
The new Dockerfile with Ubuntu 23.04 can't be used outside CI at the
moment because the `docker run ... --user $(id -u):$(id -g)` command
doesn't work with Ubuntu 23.04: Dockerfile attempts to create a user
with GID and UID that already exist. I want to concentrate on the new
release for now, so this issue will get addressed later.