WIP: fix tests
Initially I just provided operator+ which deferred the actual work to
push(). This compiled, but 5 tests failed because they were appending
a path that starts with a slash, which led to push() replacing the
existing path.

So I reverted and took the longer route: add join() which returns a new
object, and look through all the compilation errors to adjust strings in
our tests.