Remove `automatic` in `automatic && args->size() > 0` conditions
Automatic is set to true when running an operation from a macro.
It is set to false when running from a regular (bind-key) binding.
In case of bind-key bindings, no arguments can be specified, so
`args-size()` will always return 0 in those cases.
Given the above, `automatic` never has impact on the
`automatic && args->size() > 0` condition.

Value of `automatic && args->size() > 0`:
| type \ num_args | 0     | >= 1         |
| macro           | false | true         |
| bind-key        | false | not possible |