Avoid deadlock in `utils::run_program` with large inputs
Previous implementation attempted to write the entire input to the
program's stdin before reading anything from its stdout. With programs
that work in a streaming fashion (such as `sed` and `cat`), this would
deadlock on large inputs as the program is waiting to write to its
stdout while Newsboat is waiting to write to program's stdin.

This affected `filter:` feeds, and could also affect bookmarks and
rendering.

Fixed by moving the write into a separate thread as described in
https://doc.rust-lang.org/std/process/index.html#handling-io Would be
nice to do this in a single thread with non-blocking I/O, but
`std::process::ChildStdout` doesn't seem to have a non-blocking `read`.

Reported by B-| on IRC (the `filter:` variant). Kudos to Dennis van der
Schagt for the discussion and the suggested fix.