Speed up `utils::run_program` test
Instead of sleeping, wait on a conditional variable with a timeout. This
is beneficial because our C++ tests are single-threaded. I didn't bother
with the same optimization on the Rust side because there the test runs
as a separate binary and the overhead is unnoticeable.

Kudos to Dennis van der Schagt for the suggestion.