Do not kill the helper thread in `utils::run_program` test
In b02d5d10a3f5006fad46faae9c267147420bc29c I added a test that invoked
`utils::run_program()` in a separate thread and `SIGKILL`ed it at the
end of the test. This led to sporadic failures of the test suite,
because (apparently) sometimes it killed the main thread instead of the
helper one.

The reason I tried to kill the thread is twofold:

1. dropping `std::thread` attached to a running thread invokes
   `std::terminate`;
2. I didn't want to leave threads hanging around.

The first point is easily addressed by calling `.detach()` (which the
original code already did). The second point I think was overblown: the
test suite doesn't run that long, and the helper thread doesn't consume
that many resources, so it's pretty safe to leave it hanging.

This commit removes the call to `pthead_kill` so only `.detach()`
remains. It also puts all the data that the helper thread uses into
shared pointers, to avoid any use-after-free issues if a test finishes
while the thread is still running.