Rephrase slightly
Most of the time we spell out "configuration" in running text, so let's
stick to the long form here, too.

Mixing plural ("These") and singular ("last character") just sounds a
wee bit clumsy. So convert the plural form to singular, because the rest
of the explanation also uses singular.

An indefinite article makes the sentence read more fluently.

I didn't reformat to 80 columns to keep the diff clean.