CI: bump Rust to 1.80.0
This release enabled automatic checking of `cfg`s at compile-time[1], so
Clippy started warning us about the use of `#[cfg(feature
= "nightly")]`. Apparently I misunderstood how it works when I was
copying code from the human-panic crate.

While trying to make the warning go away, I also realized that the
Nightly-only code doesn't even compile, because we declare a function to
return `String` but actually return `()`.

Apparently the proper way to fix this is to add compiler auto-detection
to build.rs. This will slow down compilation a bit, for what I perceive
to be a very small benefit; after all, we got this far without that
Nightly-only API. Thus I removed the code that used the API.

1. https://blog.rust-lang.org/2024/05/06/check-cfg.html