Migrate to source-based code coverage
We used to rely on `-Zprofile`, but it was recently removed from Rust
Nightly: https://github.com/rust-lang/rust/pull/131829 This commit
switches to a stable alternative; cf.
https://github.com/mozilla/grcov/issues/1240

This change is accompanied by a change in the build environment: instead
of using Clang 18 with Rust Nightly, we now use Clang 18 with Rust 1.81.
That's the last Rust version based on LLVM 18. Previously we tried to
match LLVM versions between C++ and Rust to avoid weird coverage
results. Now, it's *required* to match them, otherwise no coverage is
gathered at all.

Fixes #2912.