Report error if OPML file is missing <opml> or <body>
User kd4ihw_home reported an issue in our IRC channel. Newsboat showed a
success message that the OPML file was successfully imported. However,
the urls file wasn't touched at all, no URLs had been added.

It turned out that the OPML file was actually invalid. The exact problem
with the OPML file wasn't confirmed, but the <body> element in the
<opml> root element must have been missing, so no data were processed.
The provided strace output did not include any write attempt to the urls
file, supporting this theory.

Now, we show a dedicated error message if there is no <body> element.
For good measure we now also abort right away if the root element is not
an <opml> element. This should improve the user experience and avoid
debugging effort.