Make RssItem almost-constant
It returns dummy data and stores next to nothing. If lots of small
strings is the core of the problem, then this commit should noticeably
help.

The results were:

    start: 21M
    reload: 98M
    reload: 109M
    reload: 109M
    reload: 109M
    enter feed
    enter item
    exit back to feedlist: 110M
    reload: 110M
    exit & start: 66M
    reload: 137M
    reload: 150M
    reload: 150M
    reload: 151M
    exit & start: 66M
    reload: 108M
    reload: 150M
    reload: 151M

It looks to me like the problem is still there: even after the first
reload that doubles the memory use, RSS continues to climb before
leveling off.