Fix cookies not written by curl 8.17+
The culprit is curl commit fd6eb8d6e77d95e71c0c55678b46173b21edd1e9:

    commit fd6eb8d6e77d95e71c0c55678b46173b21edd1e9
    Author: Daniel Stenberg <daniel@haxx.se>
    Date:   Fri Sep 19 15:59:57 2025 +0200

        cookie: avoid saving a cookie file if no transfer was done

        Because parts of the cookie loading happens on transfer start the
        in-memory cookie jar risks being incomplete and then a save might
        wrongly truncate the target file.

        Added test 1902 to verify.

        Reported-by: divinity76 on github
        Fixes #18621
        Closes #18622

     lib/cookie.c               | 12 ++++++------
     tests/data/Makefile.am     |  2 +-
     tests/data/test1902        | 43 +++++++++++++++++++++++++++++++++++++++++
     tests/libtest/Makefile.inc |  2 +-
     tests/libtest/lib1902.c    | 48 ++++++++++++++++++++++++++++++++++++++++++++++
     5 files changed, 99 insertions(+), 8 deletions(-)
     create mode 100644 tests/data/test1902
     create mode 100644 tests/libtest/lib1902.c

This doesn't play nicely with Newsboat's re-use of curl handles. After
an HTTP transfer, we reset the handle to avoid leaking the options to
the next transfer, but we also restore the cookiejar option so that the
cookies are written when `curl_easy_cleanup()` is called. The reset
makes it look like there were no transfers on the handle, so curl 8.17+
doesn't write anything to the cookiejar.

The fix is to reset the handle *before* re-using it.